Move iterable to core
