+++ src/core/__init__.py
@@ -1,3 +1,5 @@
 __all__ = [
+    "iterable",
+    "fileio",
     "gridutils"
 ]
