+++ src/core/tests/test_iterable.py
@@ -1,4 +1,4 @@
-from mathtools.iterable import take, unzip
+from core.iterable import take, unzip
 
 def test_take() -> None:
     assert take(0, range(0, 10)) == []
