+++ src/mathtools/__init__.py
@@ -9,7 +9,6 @@ from typing import overload, Any, Iterable, Union
 __all__ = [
     'combinatorics',
     'functional',
-    'iterable',
     'number_theory'
 ]
